+++ tools/emulator/device.c
@@ -80,7 +80,7 @@ uint32_t readDeviceRegister(uint32_t address)
 			{
 				struct timeval tv;
 				gettimeofday(&tv, NULL);
-				return tv.tv_sec * 1000000 + tv.tv_usec;
+				return (uint32_t)(tv.tv_sec * 1000000 + tv.tv_usec);
 			}
 
 		case 0x48:
